Update README.md
My blog offers a comprehensive analysis of Toncoin's price prediction for 2024, 2025-2030, and 2050, based on thorough research from various well-known cryptocurrency exchanges and analysts. It forecasts a bullish long-term market sentiment for Toncoin, with multiple interlinks to Benzinga for additional resources. While optimistic, the blog emphasizes that cryptocurrency is highly volatile, and no prediction can be guaranteed.